backtesting_backend/requirements.txt
# backtesting_backend/requirements.txt (Python 3.12 환경 최종 최적화)
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
aiosqlite==0.20.0
SQLAlchemy==2.0.30
pandas==2.3.3
numpy==2.2.6
pandas-ta==0.4.71b0
python-binance==1.0.19
websockets==12.0
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
httpx==0.27.0  # 필요한 버전으로 명시 (현재 0.27.0이 안정적)